Filtering internal timescaledb tables, views, and routines
Timescaledb creates many internal database objects if a postgres table
has the extension enabled. The number of these objects (especially tables)
increase drastically as more data is stored, to the point that the table
view fails to load and no tables (timescaledb or not) are viewable.

This change filters out all of these timescaledb specific internal objects
(any timescaledb schema prefixed with an underscore) which the user will
never care about. In addition to fixing the table view not loading, this
removes unnecessary clutter from views and routines.

branch:timescaledb-filtering